--- a/spec/integrations/any_fixture_spec.rb
+++ b/spec/integrations/any_fixture_spec.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require "spec_helper"
-
 describe "AnyFixture" do
   specify "it works" do
     output = run_rspec("any_fixture")