--- a/spec/integrations/factory_all_stub_spec.rb
+++ b/spec/integrations/factory_all_stub_spec.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require "spec_helper"
-
 describe "FactoryAllStub" do
   specify "it works" do
     output = run_rspec("factory_all_stub")