--- a/spec/integrations/factory_default_spec.rb
+++ b/spec/integrations/factory_default_spec.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require "spec_helper"
-
 describe "FactoryDefault" do
   specify "RSpec integration", :aggregate_failures do
     output = run_rspec("factory_default")