--- a/spec/integrations/profilers_spec.rb
+++ b/spec/integrations/profilers_spec.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require "spec_helper"
-
 PROFILERS_AVAILABLE =
   begin
     require "stackprof"