--- a/spec/integrations/sample_spec.rb
+++ b/spec/integrations/sample_spec.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require "spec_helper"
-
 describe "Tests Sampling" do
   context "RSpec integration" do
     specify "SAMPLE=2" do