--- a/spec/test_prof/any_fixture_spec.rb
+++ b/spec/test_prof/any_fixture_spec.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-require "spec_helper"
 require "test_prof/any_fixture"
 
 describe TestProf::AnyFixture, :transactional do