--- a/spec/test_prof/event_prof/profiler_spec.rb
+++ b/spec/test_prof/event_prof/profiler_spec.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require "spec_helper"
-
 describe TestProf::EventProf::Profiler do
   let(:rank_by) { :time }
   let(:top_count) { 5 }