--- a/spec/test_prof/ext/active_record_refind_spec.rb
+++ b/spec/test_prof/ext/active_record_refind_spec.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-require "spec_helper"
 require "test_prof/ext/active_record_refind"
 
 using TestProf::Ext::ActiveRecordRefind