--- a/spec/test_prof/ext/float_duration_spec.rb
+++ b/spec/test_prof/ext/float_duration_spec.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-require "spec_helper"
 require "test_prof/ext/float_duration"
 
 using TestProf::FloatDuration