--- a/spec/test_prof/ext/string_truncate_spec.rb
+++ b/spec/test_prof/ext/string_truncate_spec.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-require "spec_helper"
 require "test_prof/ext/string_truncate"
 
 using TestProf::StringTruncate