--- a/spec/test_prof/factory_prof/printers/flamegraph_spec.rb
+++ b/spec/test_prof/factory_prof/printers/flamegraph_spec.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require "spec_helper"
-
 describe TestProf::FactoryProf::Printers::Flamegraph do
   subject { described_class }