--- a/spec/test_prof/recipes/active_record_shared_connection_spec.rb
+++ b/spec/test_prof/recipes/active_record_shared_connection_spec.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-require "spec_helper"
 require "test_prof/recipes/active_record_one_love"
 
 describe TestProf::ActiveRecordSharedConnection, :transactional do