--- a/spec/test_prof/rspec_stamp/parser_spec.rb
+++ b/spec/test_prof/rspec_stamp/parser_spec.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-require "spec_helper"
 require "test_prof/rspec_stamp/parser"
 
 describe TestProf::RSpecStamp::Parser do