--- a/spec/test_prof/ruby_prof_spec.rb
+++ b/spec/test_prof/ruby_prof_spec.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require "spec_helper"
-
 describe TestProf::RubyProf do
   # Use fresh config all for every example
   after { described_class.remove_instance_variable(:@config) }