--- a/spec/test_prof/utils/sized_ordered_set_spec.rb
+++ b/spec/test_prof/utils/sized_ordered_set_spec.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-require "spec_helper"
 require "test_prof/utils/sized_ordered_set"
 
 describe TestProf::Utils::SizedOrderedSet do