--- a/spec/test_prof_spec.rb
+++ b/spec/test_prof_spec.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require "spec_helper"
-
 describe TestProf do
   describe "#artifact_path" do
     before do