### What is the purpose of this pull request?

Get rid of some boilerplate with an unnoticeable trade-off.

In addition to that:
 - it's possible to add `binding.pry` in the spec code without `require 'pry'`
 - RSpec's config is being loaded, which:
 - - adds and removes `tmp`
 - - sets execution order to `random` when running separate spec files (even those that didn't include `spec_helper` - found two issues in `let_it_be` with random order being set

More on the trade-off part below.

According to my measurements, time to run a single spec is 1.15s after,
and 0.9s before this change for `let_it_be_fixture.rb`.

When running the entire suite one by one (for each spec file RSpec
initialization is repeated to accumulate the effect), the difference is
even less noticeable:

    $ time fd _spec.rb | xargs -n 1 rspec

before:

    xargs -n 1 rspec  106.96s user 24.19s system 88% cpu 2:28.19 total

after:

    xargs -n 1 rspec  108.26s user 24.49s system 88% cpu 2:29.71 total

### What changes did you make? (overview)

Removed explicit requires of `spec_helper` and added a directive to require it from `.rspec` config file.

### Is there anything you'd like reviewers to focus on?

I might have missed some of the integration files in my benchmark. Is there anything else?

Side observation: example persistence file setting is useless:

```ruby
  config.example_status_persistence_file_path = "tmp/.rspec-status" if
    config.respond_to?(:example_status_persistence_file_path=)

  config.after(:suite) do
    FileUtils.rm_rf("tmp")
  end
```

should either be outside of `tmp`, or `tmp` should be kept after spec execution.

### Checklist

- [-] I've added tests for this change
- [-] I've added a Changelog entry
- [-] I've updated a documentation